address the cudnn build warnings (#6)
* fix warning: unused variable ‘devcount’ [-Wunused-variable] in backend_cudnn.cc.
* fix warning: format ‘%f’ expects argument of type ‘double’, but argument has type ‘int’.
* fix warning: control reaches end of non-void function [-Wreturn-type].
* fix warning: comparison of integer expressions of different signedness: ‘size_t’ {aka ‘long unsigned int’} and ‘int’ [-Wsign-compare].

+++ src/executable/naive_conv.h
@@ -72,7 +72,7 @@ private:
 template<class blockwise_t, class threadwise_t, class... args_t>
 void naive_conv_blockwise_in_parallel(threadwise_t thread_func, args_t... args)
 {
-    int num_threads = std::thread::hardware_concurrency();
+    size_t num_threads = std::thread::hardware_concurrency();
     std::vector<std::thread> threads(num_threads);
 
     for(size_t tid = 0; tid < num_threads; tid++)
